jobs: factorise common job logic into BaseJob (bug 1983208)
