+++ src/lando/headless_api/models/automation_job.py
@@ -1,54 +1,27 @@
 from contextlib import contextmanager
 from datetime import datetime
-from typing import Any, Iterable, Optional, Self
+from typing import Any, Iterable, Optional
 
 from django.conf import settings
 from django.db import models
+from django.db.models import QuerySet
 from django.utils.translation import gettext_lazy
 
-from lando.main.models.base import BaseModel
-from lando.main.models.landing_job import (
-    JobAction,
+from lando.main.models import (
+    BaseJob,
+    BaseModel,
     JobStatus,
+    Repo,
 )
-from lando.main.models.repo import Repo
 
 
-class AutomationJob(BaseModel):
+class AutomationJob(BaseJob):
     """Represent an automation job request through the headless API.
 
     This job is executed by the automation worker, where the set of associated
     `AutomationAction` entires are retrieved and applied to the target repo locally
     before pushing.
     """
-
-    # Current status of the job.
-    status = models.CharField(
-        max_length=32,
-        choices=JobStatus,
-        default=None,
-    )
-
-    # Email of the user who created the automation job.
-    requester_email = models.CharField(blank=True, default="", max_length=255)
-
-    # Identifier for the most descendent commit created by this landing.
-    landed_commit_id = models.TextField(blank=True, default="")
-
-    # Number of attempts made to complete the job.
-    attempts = models.IntegerField(default=0)
-
-    # Priority of the job. Higher values are processed first.
-    priority = models.IntegerField(default=0)
-
-    # Duration of job from start to finish
-    duration_seconds = models.IntegerField(default=0)
-
-    # Reference to the target repo.
-    target_repo = models.ForeignKey(Repo, on_delete=models.SET_NULL, null=True)
-
-    # Text describing errors when status != LANDED.
-    error = models.TextField(default="", blank=True)
 
     # Name of RelBranch to push changes to.
     relbranch_name = models.CharField(null=True, blank=True)
@@ -72,74 +45,27 @@ class AutomationJob(BaseModel):
 
     def to_api_status(self) -> dict[str, Any]:
         """Return the job details as API status JSON."""
-        return {
-            "job_id": self.id,
-            "status_url": f"{settings.SITE_URL}/api/job/{self.id}",
-            "message": f"Job is in the {self.status} state.",
-            "created_at": self.created_at,
-            "status": self.status,
-            "error": self.error,
-        }
+        job_dict = self.to_dict()
 
-    def transition_status(
-        self,
-        action: JobAction,
-        **kwargs,
-    ):
-        """Change the status and other applicable fields according to actions.
+        job_dict["job_id"] = job_dict.pop("id")
+        job_dict["message"] = f"Job is in the {self.status} state."
+        job_dict["status_url"] = f"{settings.SITE_URL}/api/job/{self.id}"
 
-        Args:
-            action (JobAction): the action to take, e.g. "land" or "fail"
-            **kwargs:
-                Additional arguments required by each action, e.g. `message` or
-                `commit_id`.
-        """
-        actions = {
-            JobAction.LAND: {
-                "required_params": ["commit_id"],
-                "status": JobStatus.LANDED,
-            },
-            JobAction.FAIL: {
-                "required_params": ["message"],
-                "status": JobStatus.FAILED,
-            },
-            JobAction.DEFER: {
-                "required_params": ["message"],
-                "status": JobStatus.DEFERRED,
-            },
-            JobAction.CANCEL: {
-                "required_params": [],
-                "status": JobStatus.CANCELLED,
-            },
-        }
-
-        if action not in actions:
-            raise ValueError(f"{action} is not a valid action")
-
-        required_params = actions[action]["required_params"]
-        if sorted(required_params) != sorted(kwargs.keys()):
-            missing_params = required_params - kwargs.keys()
-            raise ValueError(f"Missing {missing_params} params")
-
-        self.status = actions[action]["status"]
-
-        if action in (JobAction.FAIL, JobAction.DEFER):
-            self.error = kwargs["message"]
-
-        if action == JobAction.LAND:
-            self.landed_commit_id = kwargs["commit_id"]
-
-        self.save()
+        return job_dict
 
     @classmethod
-    def next_job(cls, repositories: Optional[Iterable[str]] = None) -> Self:
-        """Return the next automation job."""
-        return (
-            cls.objects.filter(status__in=(JobStatus.SUBMITTED, JobStatus.DEFERRED))
-            .filter(target_repo__in=repositories)
-            .order_by("-priority", "created_at")
-            .select_for_update()
-        )
+    def job_queue_query(
+        cls, repositories: Optional[Iterable[str]] = None, **kwargs
+    ) -> QuerySet:
+        """Return a query which selects the queued jobs."""
+        # XXX: If we can include IN_PROGRESS jobs, and sort by status_order, this method
+        # can be deleted in favour of the parent class's.
+        q = cls.objects.filter(status__in=(JobStatus.SUBMITTED, JobStatus.DEFERRED))
+        if repositories:
+            q = q.filter(target_repo__in=repositories)
+
+        q = q.order_by("-priority", "created_at")
+        return q
 
     def resolve_push_target_from_relbranch(self, repo: Repo) -> tuple[str | None, str]:
         """Return (target_cset, push_target) tuple for the `RelBranchSpecifier` if required."""
