+++ src/lando/main/models/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 from lando.main.models.base import *
 from lando.main.models.configuration import *
+from lando.main.models.jobs import *
 from lando.main.models.landing_job import *
 from lando.main.models.profile import *
 from lando.main.models.revision import *
